+++ packages/discovery-react-components/src/components/DocumentPreview/components/Highlight/__tests__/TableHighlight.test.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { act, render, waitForElement } from '@testing-library/react';
 import Highlight from '../Highlight';
-import docJson from '../../../__fixtures__/Art Effects Koya Creative Base TSA 2008.pdf.json';
+import docJson from 'components/DocumentPreview/__fixtures__/Art Effects Koya Creative Base TSA 2008.pdf.json';
 
 const tableResult = {
   table_id: '558ada041262d5b0aa02a05429d798c7',