+++ packages/discovery-react-components/src/components/SearchResults/__tests__/SearchResults.test.tsx
@@ -6,8 +6,8 @@ import {
   SearchContextIFC,
   SearchApiIFC,
   searchResponseStoreDefaults
-} from '../../DiscoverySearch/DiscoverySearch';
-import { wrapWithContext } from '../../../utils/testingUtils';
+} from 'components/DiscoverySearch/DiscoverySearch';
+import { wrapWithContext } from 'utils/testingUtils';
 import { SearchResults, SearchResultsProps } from '../SearchResults';
 import { getByText as domGetByText } from '@testing-library/dom';